+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### To be released
 - Support Icons for Input (https://github.com/tony-dinh/react-toolkit/pull/28)
 - Tabs Component (https://github.com/tony-dinh/react-toolkit/pull/29)
+- Switch to React 16.0.0
+- Form Component add `validateOnUpdate` prop validation control on input updates (https://github.com/tony-dinh/react-toolkit/pull/30)
 
 ### v0.2.0
 - Segmented Bar Graph Component (https://github.com/tony-dinh/react-toolkit/pull/24)